+++ java-vendingmachine-precourse/src/main/java/vendingmachine/IVendingMachine.java
@@ -1,0 +1,19 @@
+package vendingmachine;
+
+import vendingmachine.Coin.Pocket;
+
+public interface IVendingMachine{
+    void depositMachineBalance(Pocket pocket);
+
+    void addMachineStock(Stock stock);
+
+    void depositUserBalance(int balance);
+
+    int getUserBalance();
+
+    boolean canUserPurchase();
+
+    void purchaseProductByName(String name);
+
+    Coin.Pocket withdrawUserChange();
+}
